fix: Update release-please configuration for Terraform and Terragrunt
The changes in this commit update the release-please configuration for the `infra/terraform` and `infra/terragrunt` packages. The release-type is changed from `terraform` to `simple`, and the `bump-minor-pre-major` and `bump-patch-for-minor-pre-major` options are retained.

Additionally, the configuration for the `tools/infractl` package has been removed, as it is no longer needed.

The corresponding change has been made in the `.github/workflows/release.yml` file, where the `release-type` has been updated to `simple`.