fix-release-please (#8)
* feat: Update release-please configuration

Improve the release-please configuration by:
- Adding more detailed changelog sections
- Enabling extra file updates (README.md)
- Configuring the release type to be "simple"
- Removing the manifest file as it's no longer needed

* fix: Update release workflow to use simple release type

The release workflow is updated to use the `simple` release type instead of the `main` release type. This change ensures that the release process is more straightforward and consistent with the project's release strategy.

* ci: Remove workflow_dispatch trigger from release workflow

The `workflow_dispatch` trigger has been removed from the release workflow. This change ensures that the release workflow is only triggered on pushes to the `main` branch and any `[0-9]+.[0-9]+.x` branches, without the ability to manually trigger the workflow.