feat/add-nix-shell (#20)
* feat: Add Nix development shell configuration

Introduce a Nix development shell configuration for the Terraform Registry Module Template project. The shell includes a set of tools and utilities commonly used for Terraform development, such as:

- Go toolchain (go, go-tools, golangci-lint)
- Terraform and related tools (terraform, terraform-ls, tflint, opentofu, terraform-docs, terragrunt)
- General development and utility tools (just, git, bash)
- YAML tools (yamllint, yamlfmt)
- Shell scripting tools (shellcheck)
- direnv for environment management

The development shell also includes a shell hook that prints the versions of the key tools (Go, Terraform, OpenTofu, Terragrunt) when the shell is entered.

Additionally, a new `dev` target is added to the `justfile` to start the Nix development shell.

* feat: Improve development environment with a new shell and tool versions

- Add a minimal default shell with essential tools for fast startup
- Add a full development shell with extended tools
- Extract tool version information to a separate script for easy access
- Update the description in the flake.nix file to better reflect the purpose of the environment