feat: Add local git base URL to generator modules
The changes introduce a new local variable `git_base_url_local` to the generator module configurations (`dni_generator.hcl`, `name_generator.hcl`, `lastname_generator.hcl`, and `age_generator.hcl`). This variable is set to the value of `local.cfg.locals.cfg_git.git_base_urls.local`, which allows for a separate local git base URL to be used for the Terraform modules.

This change provides more flexibility in managing the source of the Terraform modules, allowing for a distinct local git repository to be used in addition to the main remote repository.