feat: Improve Terragrunt command execution with auto-approve for apply and destroy (#22)
* feat: Improve Terragrunt command execution with auto-approve for apply and destroy

This commit enhances the `tg-run` recipe in the `justfile` to automatically include the `--auto-approve` flag when running the `apply` or `destroy` Terragrunt commands. This change simplifies the execution of these commands, reducing the need for manual confirmation during the deployment process.

* feat: Add local git base URL to generator modules

The changes introduce a new local variable `git_base_url_local` to the generator module configurations (`dni_generator.hcl`, `name_generator.hcl`, `lastname_generator.hcl`, and `age_generator.hcl`). This variable is set to the value of `local.cfg.locals.cfg_git.git_base_urls.local`, which allows for a separate local git base URL to be used for the Terraform modules.

This change provides more flexibility in managing the source of the Terraform modules, allowing for a distinct local git repository to be used in addition to the main remote repository.

* feat: Add local module path configuration for testing

The changes in this commit add a TODO comment to the terragrunt.hcl files for the age, name, dni, and lastname generators. The TODO comment suggests switching the `tf_module_local_path` to use a local module path when testing locally, and using the remote module set in the `_shared/_units/<my_unit>.hcl` file if the local path is empty.

This change allows for more flexible module sourcing, enabling developers to easily switch between local development and testing, and using the remote module for production deployments. This improves the development workflow and ensures consistent module referencing across the infrastructure.

* feat: Improve README.md with detailed infrastructure unit configuration details

- Add a section explaining the architecture principles and core design concepts
- Describe the module source management approach, including the benefits of the centralized configuration
- Explain the intelligent tagging system and the advantages of the multi-layered tagging strategy
- Outline the dynamic configuration loading mechanism and the benefits of the hierarchical configuration control
- Discuss the dependency management capabilities and the advantages of the robust dependency resolution and mocking
- Provide guidance on the integration mechanism and the standardized approach for including shared units
- Summarize the best practices for module management, configuration design, and dependency handling
- Include a section on troubleshooting common resolution strategies for module source verification, configuration loading diagnostics, and dependency resolution